01AICodeSnippets/07-google-agentic-sdk/requirements.txt
# Core dependencies
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2

# Utilities
requests>=2.26.0
python-multipart>=0.0.5 